simplified module structure by running handlers as txtminebot
